requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
python-binance>=1.0.16
yfinance>=0.2.0

# Machine Learning & Optimization
scikit-learn>=1.1.0
joblib>=1.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Progress bars
tqdm>=4.64.0

# Environment management
python-dotenv>=0.19.0

# Technical indicators (substitui TA-Lib)
ta>=0.10.0

# Additional utilities
requests>=2.28.0
psutil>=5.9.0